flask-pdf-to-word-app/requirements.txt
Flask
pdf2image
pytesseract
python-docx
Werkzeug
Pillow
pdf2docx